Move TimeSignature class to Data Access project
Move TimeSignature class to Data Access project
