+++ Core/MTDataAccess/MTDataAccess.csproj
@@ -65,6 +65,7 @@
     <Compile Include="DB.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SQL.cs" />
+    <Compile Include="TimeSignature.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
