+++ Multitracks_Api/Multitracks_Api.csproj
@@ -152,9 +152,9 @@
     <Compile Include="Models\Artist.cs" />
     <Compile Include="Models\ResponseModel.cs" />
     <Compile Include="Models\Song.cs" />
-    <Compile Include="Models\TimeSignature.cs" />
-    <Compile Include="obj\Debug\.NETFramework,Version=v4.8.AssemblyAttributes.cs" />
+    <!-- TimeSignature.cs removed as it's now in MTDataAccess -->
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="obj\Debug\.NETFramework,Version=v4.8.AssemblyAttributes.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="bin\Multitracks_Api.dll.config" />
